+++ src/ui/charts/CDFChart.tsx
@@ -47,7 +47,7 @@ export function CDFChart({
         <ResponsiveContainer width="100%" height={300}>
           <LineChart
             data={data}
-            margin={{ top: 10, right: 30, left: 0, bottom: 0 }}
+            margin={{ top: 10, right: 30, left: 0, bottom: 20 }}
           >
             <CartesianGrid strokeDasharray="3 3" />
             <XAxis